+++ budget/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- The transaction list filters now include a **Reconciled** filter (All / Reconciled / Not reconciled), so you can quickly see which transactions still need reconciling ([#301](https://github.com/otherworld-dev/Budget/discussions/301))
+
 ## [2.33.1] - 2026-06-18
 
 ### Fixed